% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_sp_curves.R
\name{build_sp_curves}
\alias{build_sp_curves}
\title{Build surplus production curves}
\usage{
build_sp_curves(output, cov_vals)
}
\arguments{
\item{output}{SP model output}

\item{cov_vals}{Covariate models to build curves for}
}
\value{
A dataframe with SP curves.
}
\description{
This function builds surplus production curves.
}
