% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_sr_curves.R
\name{build_sr_curves}
\alias{build_sr_curves}
\title{Build stock recruit curves}
\usage{
build_sr_curves(output, cov_vals)
}
\arguments{
\item{output}{SR model output}

\item{cov_vals}{Covariate models to build curves for}
}
\value{
A dataframe with SR curves.
}
\description{
This function builds stock recruit curves.
}
