% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_msy.R
\name{calc_msy}
\alias{calc_msy}
\title{Calculate covariate-linked MSY over time}
\usage{
calc_msy(output, cov_df)
}
\arguments{
\item{output}{Output of model fit}

\item{cov_df}{Data frame containing: stockid, year, cov_col}
}
\description{
This function calculates MSY.
}
