% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_models.R
\name{compare_models}
\alias{compare_models}
\title{Compare models with AIC}
\usage{
compare_models(models, names)
}
\arguments{
\item{models}{List with models}

\item{names}{Names of models}
}
\value{
A table with AIC comparison
}
\description{
This function compares models with AIC.
}
