% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields_data.R
\name{fields_data}
\alias{fields_data}
\title{Types of acceptable fields for splink searching}
\usage{
fields_data()
}
\value{
data.frame with fields of setup_list
}
\description{
Types of acceptable fields for splink searching
}
\details{
fields_data returns a data.frame of the fields that are supported and the respective types. Searchable are free fields and the modifiers are close fields to modify one or more specific searchable field.
}
