% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sp.R
\name{fit_sp}
\alias{fit_sp}
\title{Fit surplus production model using maximum likelihood estimation}
\usage{
fit_sp(data, p = 1)
}
\arguments{
\item{data}{A prepped data frame with the following required columns: stockid, year, biomass_scaled, sp_scaled}

\item{p}{The shape parameter}
}
\value{
A list containing: (1) the data; (2) the model fit; and (3) the standard error report.
}
\description{
This function fits a surplus production model using maximum likelihood estimation in TMB.
}
