% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sp_linked_random.R
\name{fit_sp_linked_random}
\alias{fit_sp_linked_random}
\title{Fit surplus production model using maximum likelihood estimation}
\usage{
fit_sp_linked_random(data, b_col, sp_col, p = 1, cov_col)
}
\arguments{
\item{data}{Data}

\item{b_col}{Name of biomass column}

\item{sp_col}{Name of surplus production column}

\item{p}{Shape parameter}

\item{cov_col}{Name of covariate column}
}
\value{
A list containing: (1) the data; (2) the model fit; and (3) the standard error report.
}
\description{
This function fits a surplus production model using maximum likelihood estimation in TMB.
}
\examples{
data <- splink::ram_ne
output <- splink::fit_sp_linked_random(data=data, b_col="tb_scaled", sp_col="sp_scaled", p=1, cov_col="sst_c_scaled")
results <- splink::get_results(output)
splink::plot_results(results)
splink::plot_thetas(results)
splink::plot_fits_sp_linked(output, b_col="tb_scaled", sp_col="sp_scaled", cov_col="sst_c_scaled")
}
