% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sp_nls.R
\name{fit_sp_nls}
\alias{fit_sp_nls}
\title{Fit surplus production model using NLS estimation}
\usage{
fit_sp_nls(data, p = 1)
}
\arguments{
\item{data}{A prepped data frame with the following required columns: stockid, year, biomass_scaled, sp_scaled}

\item{p}{The shape parameter}
}
\value{
A data frame containing the point estimate fits for r and K
}
\description{
This function fits a surplus production model using NLS estimation.
}
