% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sr.R
\name{fit_sr}
\alias{fit_sr}
\title{Fit stock recruit model using maximum likelihood estimation}
\usage{
fit_sr(data, b_col, r_col, type, cov_col = NULL, cov_effect = NULL)
}
\arguments{
\item{data}{A data frame}

\item{b_col}{Name of the column with biomass}

\item{r_col}{Name of the column with recruitment}

\item{type}{Stock recruit type: "ricker" or "bev-holt"}

\item{cov_col}{Name of the column with the covariate}
}
\value{
A list containing: (1) the data; (2) the model fit; and (3) the standard error report.
}
\description{
This function fits a Ricker or Beverton-Holt stock recruitment model using maximum likelihood estimation in TMB.
}
\examples{
data <- splink::ram_wc
output <- splink::fit_sr(data=data, b_col="b_scaled", r_col="r_scaled", type="ricker")
results <- splink::get_results(output)
splink::plot_results(results)
splink::plot_fits_sr(output, b_col="b_scaled", r_col="r_scaled", plotdir="~/Desktop", plotname="test.pdf")
}
