% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{download dos dados}
\usage{
get_data(list_data)
}
\arguments{
\item{list_data}{list. Object returned by \code{setup_list()}, or a named list with parameters to request download.}
}
\value{
list
}
\description{
faz download dos dados do splink com base na requisição
}
\details{
https://api.splink.org.br/
}
\examples{
get_data(list_data = list(Scientificname = c("Manilkara maxima"),
                          MaxRecords = 5))
\dontrun{
get_data(list_data = list(ScientificName = c("Rauvolfia sellowii", "Cantinoa althaeifolia"),
                          StateProvince = c("São Paulo","Rio de Janeiro","Pernambuco"),
                          MaxRecords = 5,
                          Model = "Coords"))
                          }

}
