% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params.R
\name{get_results}
\alias{get_results}
\title{Fetches the parameter estimates and confidence intervals from a TMB model fit}
\usage{
get_results(output)
}
\arguments{
\item{output}{Output from function fitting biomass dynamics model}
}
\value{
A data frame with the point estimates, standard error, and 95% confidence interval from model fit
}
\description{
This function fetches the parameter estimates and confidence intervals from a TMB model fit.
}
