% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fits_nls.R
\name{plot_fits_nls}
\alias{plot_fits_nls}
\title{Plot surplus production model fits}
\usage{
plot_fits_nls(output, plotdir = getwd(), plotname = "NLS_SP_fits.pdf")
}
\arguments{
\item{output}{SP model output}

\item{plotdir}{Path for plot export}

\item{plotname}{Plot name (must end with ".pdf)}
}
\value{
PDF with production model fits
}
\description{
This function plots surplus production model fits and exports to a PDF.
}
