% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_sp_curves.R, R/plot_fits_sp_linked.R
\name{plot_fits_sp_linked}
\alias{plot_fits_sp_linked}
\title{Build surplus production curves}
\usage{
plot_fits_sp_linked(
  output,
  b_col,
  sp_col,
  cov_col,
  plotdir = getwd(),
  plotname = "TMB_fits.pdf"
)

plot_fits_sp_linked(
  output,
  b_col,
  sp_col,
  cov_col,
  plotdir = getwd(),
  plotname = "TMB_fits.pdf"
)
}
\arguments{
\item{output}{SP model output}

\item{plotdir}{Path for plot export}

\item{plotname}{Plot name (must end with ".pdf)}
}
\value{
A dataframe with SP curves.

PDF with production model fits
}
\description{
This function builds surplus production curves.

This function plots covariate-linked surplus production model fits and exports to a PDF.
}
