% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fits_sr_linked.R
\name{plot_fits_sr_linked}
\alias{plot_fits_sr_linked}
\title{Plot covariate-linked stock recruit fits}
\usage{
plot_fits_sr_linked(output, plotdir = getwd(), plotname = "TMB_fits.pdf")
}
\arguments{
\item{output}{SP model output}

\item{plotdir}{Path for plot export}

\item{plotname}{Plot name (must end with ".pdf)}
}
\value{
PDF with production model fits
}
\description{
This function plots stock recruit model fits and exports to a PDF.
}
