% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_raw.R
\name{plot_raw}
\alias{plot_raw}
\title{Plot sample of prepped surplus production data}
\usage{
plot_raw(data, n = NULL)
}
\arguments{
\item{data}{Prepped data}

\item{n}{Number of stocks to plot data for}
}
\value{
Plot of SP curves for N stocks
}
\description{
This function plots a sample of prepped surplus production data.
}
