% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_results.R
\name{plot_results}
\alias{plot_results}
\title{Plot parameter estimates and confidence intervals}
\usage{
plot_results(results)
}
\arguments{
\item{results}{Results data frame}
}
\value{
Plot of parameter estimates and confidence intervals
}
\description{
This function plots point estimates and confidence intervals for parameters in a covariate-linked biomass dynamic model fit. The plot shows histograms of the point estimates and spline plots showing the point estimates and their 95% confidence intervals.
}
