% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_thetas.R
\name{plot_thetas}
\alias{plot_thetas}
\title{Plot theta estimates and confidence intervals}
\usage{
plot_thetas(results)
}
\arguments{
\item{results}{Results data frame}
}
\value{
Plot of theta estimates and confidence intervals
}
\description{
This function plots point estimates and confidence intervals for the covriate influence parameter (theta) in a covariate-linked model fit. The spline plot shows the point estimates and their 95% confidence intervals.
}
