% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data}
\alias{prep_data}
\title{Prepare data for fitting biomass dynamics models}
\usage{
prep_data(data)
}
\arguments{
\item{data}{Raw data}
}
\value{
Prepped data - filtered, scaled
}
\description{
Prepare data.
}
