% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ram}
\alias{ram}
\title{RAM biomass and catch time series}
\format{
A data frame with the following attributes::
\describe{
  \item{stockid}{Stock id}
  \item{year}{Year}
  \item{biomass_units}{Biomass units}
  \item{biomass}{Biomass}
  \item{catch_units}{Catch units}
  \item{catch}{Catch}
}
}
\usage{
ram
}
\description{
RAM biomass and catch time series.
}
\keyword{datasets}
