% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ram_ne}
\alias{ram_ne}
\title{New England fish stock data}
\format{
A data frame with the following attributes::
\describe{
  \item{stockid}{Stock id}
  \item{year}{Year}
  \item{catch_mt}{Catch, in metric tons}
  \item{tb_mt}{Total biomass, in metric tons}
  \item{sp_mt}{Surplus production, in metric tons}
  \item{sst_c}{Sea surface temperature, in degrees Celcius}
  \item{tb_scaled}{Total biomass, scaled to maximum biomass}
  \item{sp_scaled}{Surplus production, scaled to maximum biomass}
  \item{sst_c_scaled}{Sea surface temperature, centered}
}
}
\usage{
ram_ne
}
\description{
Biomass, production, and temperature time series for 25 fish stocks in New England.
}
\keyword{datasets}
