% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fields_data.R
\name{fields_data}
\alias{fields_data}
\title{Types of searchable fields}
\usage{
fields_data()
}
\value{
data.frame with fields of setup_list
}
\description{
searchable fields
}
\details{
fields_data returns a matrix of the searchable fields that are supported.
}
