% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sp.R
\name{fit_sp}
\alias{fit_sp}
\title{Fit surplus production model using maximum likelihood estimation}
\usage{
fit_sp(data, b_col, sp_col, p = 1, cov_col = NULL, cov_effect = NULL)
}
\arguments{
\item{data}{Data}

\item{b_col}{Name of biomass column}

\item{sp_col}{Name of surplus production column}

\item{p}{Shape parameter}

\item{cov_col}{Name of covariate column}

\item{cov_effect}{Type of covariate effect: "fixed" or "random"}
}
\value{
A list containing: (1) the data; (2) the model fit; and (3) the standard error report.
}
\description{
This function fits a surplus production model using maximum likelihood estimation in TMB.
}
\examples{
data <- splink::ram_ne
output <- splink::fit_sp(data=data, b_col="tb_scaled", sp_col="sp_scaled", p=0.2)
results <- splink::get_results(output)
splink::plot_results(results)
splink::plot_fits(output)
}
