% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sr.R
\name{fit_sr}
\alias{fit_sr}
\title{Fit stock recruit model using maximum likelihood estimation}
\usage{
fit_sr(data, b_col, r_col, type)
}
\arguments{
\item{data}{A data frame}

\item{type}{"ricker" or "bev-holt"}
}
\value{
A list containing: (1) the data; (2) the model fit; and (3) the standard error report.
}
\description{
This function fits a Ricker or Beverton-Holt stock recruitment model using maximum likelihood estimation in TMB.
}
\examples{
data <- splink::ram_wc
output <- splink::fit_sr(data=data, b_col="b_scaled", r_col="r_scaled", type="ricker")
results <- splink::get_results(output)
splink::plot_results(results)
splink::plot_fits_sr(output, plotdir="~/Desktop", plotname="test.pdf")
}
