% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Search for speciesLink registers}
\usage{
get_data(list_data)
}
\arguments{
\item{list_data}{list. Object returned by \code{setup_list()}, or a named list with parameters to request download. To see the accepted parameters see type fields_data().}
}
\value{
data.frame containing the records based on the request.
}
\description{
Search for speciesLink registers based on the list.
}
\details{
This function is absed on CRIA speciesLink API v.0.1 beta, for more details see \url{https://api.splink.org.br/}.
}
\examples{
get_data(list_data = list(Scientificname = c("Manilkara maxima"),
                          MaxRecords = 5))
\dontrun{
get_data(list_data = list(ScientificName = c("Rauvolfia sellowii", "Cantinoa althaeifolia"),
                          StateProvince = c("São Paulo","Rio de Janeiro","Pernambuco"),
                          MaxRecords = 5,
                          Model = "Coords"))
                          }

}
