% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fits.R
\name{plot_fits}
\alias{plot_fits}
\title{Plot surplus production model fits}
\usage{
plot_fits(
  output,
  b_col,
  sp_col = NULL,
  r_col = NULL,
  cov_col = NULL,
  plotdir = getwd(),
  plotname = "TMB_fits.pdf"
)
}
\arguments{
\item{output}{SP model output}

\item{b_col}{Biomass column}

\item{sp_col}{Production column}

\item{r_col}{Recruitment column}

\item{cov_col}{Covariate column}

\item{plotdir}{Path for plot export}

\item{plotname}{Plot name (must end with ".pdf)}
}
\value{
PDF with production model fits
}
\description{
This function plots surplus production model fits and exports to a PDF.
}
