% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_list.R
\name{setup_list}
\alias{setup_list}
\title{criar lista de espécies}
\usage{
setup_list(
  Barcode = NULL,
  BasisOfRecord = NULL,
  InstitutionCode = NULL,
  CollectionCode = NULL,
  CatalogNumber = NULL,
  Collector = NULL,
  CollectorNumber = NULL,
  YearCollected = NULL,
  IdentifiedBy = NULL,
  YearIdentified = NULL,
  Kingdom = NULL,
  Phylum = NULL,
  Class = NULL,
  Order = NULL,
  Family = NULL,
  Genus = NULL,
  ScientificName = NULL,
  TypeStatus = NULL,
  Country = NULL,
  StateProvince = NULL,
  County = NULL,
  Locality = NULL,
  CoordinatesQuality = NULL,
  Format = NULL,
  Separator = NULL,
  MaxRecords = NULL,
  Model = NULL,
  Phonetic = NULL,
  RedList = NULL,
  Scope = NULL,
  Coordinates = NULL,
  Images = NULL,
  Synonyms = NULL,
  Typus = NULL,
  ShowEmptyValues = NULL,
  fieldsCase = NULL,
  Summary = NULL
)
}
\arguments{
\item{ScientificName}{vetor com os nomes das espécies}
}
\value{
list
}
\description{
cria uma lista para busca
}
\examples{
setup_list(ScientificName = "Manilkara maxima")

\dontrun{
setup_list(ScientificName = c("Rauvolfia sellowii", "Cantinoa althaeifolia"),
           StateProvince = c("São Paulo","Rio de Janeiro","Pernambuco"),
           MaxRecords = 5,
           Model = "Coords")
           }


}
