% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_list.R
\name{setup_list}
\alias{setup_list}
\title{criar lista de espécies}
\usage{
setup_list(
  Barcode = NULL,
  BasisOfRecord = NULL,
  InstitutionCode = NULL,
  CollectionCode = NULL,
  CatalogNumber = NULL,
  Collector = NULL,
  CollectorNumber = NULL,
  YearCollected = NULL,
  IdentifiedBy = NULL,
  YearIdentified = NULL,
  Kingdom = NULL,
  Phylum = NULL,
  Class = NULL,
  Order = NULL,
  Family = NULL,
  Genus = NULL,
  ScientificName = NULL,
  TypeStatus = NULL,
  Country = NULL,
  StateProvince = NULL,
  County = NULL,
  Locality = NULL,
  CoordinatesQuality = "any",
  Format = NULL,
  Separator = NULL,
  MaxRecords = NULL,
  Model = NULL,
  Phonetic = NULL,
  RedList = NULL,
  Scope = NULL,
  Coordinates = NULL,
  Images = NULL,
  Synonyms = NULL,
  Typus = NULL,
  ShowEmptyValues = NULL,
  fieldsCase = NULL,
  Summary = NULL
)
}
\arguments{
\item{Barcode}{.}

\item{BasisOfRecord}{.}

\item{InstitutionCode}{string. Accepts institution acronym. Ex: "UNICAMP", "Fiocruz".}

\item{CollectionCode}{string. collection acronym. Ex: "FIOCRUZ-CEIOC", "UEC", "HUEFS".}

\item{CatalogNumber}{numeric.     Number of specimen in the collection. Also accepts 'embranco' or "naobranco". "embranco" returns only specimens without catalog number and "naobranco" resturns only specimens with catalog number.}

\item{Collector}{string. Collector name. Ex:    "Kock", "Almeida F".}

\item{CollectorNumber}{numeric.}

\item{YearCollected}{numeric. Year of specimen was collected with four-digits.}

\item{IdentifiedBy}{string. Who identified the species of specimen. Ex: "Abreu MC"}

\item{YearIdentified}{numeric. Year of specimen was identified with four-digits.}

\item{Kingdom}{.}

\item{Phylum}{.}

\item{Class}{.}

\item{Order}{.}

\item{Family}{.}

\item{Genus}{.}

\item{ScientificName}{vector. Accepts a vector of names.}

\item{TypeStatus}{.}

\item{Country}{.}

\item{StateProvince}{.}

\item{County}{.}

\item{Locality}{.}

\item{CoordinatesQuality}{Refere-se à qualidade das coordenadas. "Good" são registros que tiveram as coordenadas checadas, "Bad" não tiveram as coordenadas checadas e "any" (default) retorna coordenadas com qualquer qualidade.}

\item{Format}{JSON | XML | CSV | TAB ## MANTER?}

\item{Separator}{especifica qual o seprador de colunas "comma" (default) ou semicolon. Este argumento é valid only for Format = CSV. ## MANTER?}

\item{MaxRecords}{numeric. Especificar o número máximo de registros devem ser retornados # POR ESPÉCIE? ##. O default é retornar todos os registros. # criar um check para n > 0.     all records}

\item{Model}{DwC | modelling}

\item{Phonetic}{Premite uma busca fonética. Ou seja, permite que o sistema ignore algumas diferenças de grafia em nomes científicos. Por exemplo: I ou Y e letras duplas. Este argumento afeta apenas as buscas em \code{filo}, \code{classe}, \code{ordem}, \code{família} e \verb{nome científico}.
Yes | No      affects only taxonomic fields: class, phylum, order, family, genus, scientificname}

\item{RedList}{Yes | No      no check}

\item{Scope}{plants, animals, microrganisms,fossils     all groups}

\item{Coordinates}{Yes | No | Original | Automatic | Blocked      no check}

\item{Images}{Yes | No | Live | Polen | Wood      no check}

\item{Synonyms}{Procura também por sinônimos definidos em alguns dicionários. Os valores aceitos são "sp2000", "flora2020", "MycoBank", "AlgaeBase", "DSMZ", "Moure" e/ou NULL (default) caso não queira realizar busca de sinônimos. Para mais infromações, veja a seção detalhes. ## IMPLEMENTAR}

\item{Typus}{Se TRUE, seleciona apenas registros que sejam tipos nomenclaturais. FALSE retorna apenas registros que não são tipos nomemclaturais. NULL é o default e retorna independente de ser tipo. ## Yes | No \emph{no check}}

\item{ShowEmptyValues}{Yes | \emph{No}}

\item{fieldsCase}{Lower | Upper | \emph{Mixed}}

\item{Summary}{Yes | \emph{No}}
}
\value{
list
}
\description{
cria uma lista para busca
}
\details{
For more information about fields see \url{https://api.splink.org.br/} or \url{http://www.splink.org.br/}.
Nos campos \strong{buscadores} são aceitos valores 'embranco' ou 'naobranco'.
}
\examples{
setup_list(ScientificName = "Manilkara maxima")

\dontrun{
setup_list(ScientificName = c("Rauvolfia sellowii", "Cantinoa althaeifolia"),
           StateProvince = c("São Paulo","Rio de Janeiro","Pernambuco"),
           MaxRecords = 5,
           Model = "Coords")
           }

}
