% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_list.R
\name{setup_list}
\alias{setup_list}
\title{Create a list object to request download of species register}
\usage{
setup_list(
  Barcode = NULL,
  BasisOfRecord = NULL,
  CollectionCode = NULL,
  CatalogNumber = NULL,
  Collector = NULL,
  CollectorNumber = NULL,
  YearCollected = NULL,
  IdentifiedBy = NULL,
  YearIdentified = NULL,
  Kingdom = NULL,
  Phylum = NULL,
  Class = NULL,
  Order = NULL,
  Family = NULL,
  Genus = NULL,
  ScientificName = NULL,
  TypeStatus = NULL,
  Country = NULL,
  StateProvince = NULL,
  County = NULL,
  Locality = NULL,
  MaxRecords = NULL,
  all.data = FALSE,
  Phonetic = FALSE,
  RedList = NULL,
  Scope = NULL,
  Coordinates = NULL,
  Images = NULL,
  Synonyms = NULL,
  Typus = NULL
)
}
\arguments{
\item{Barcode}{Barcode of specimen in a collection. Ex: NY00000001 or FPS00257.}

\item{BasisOfRecord}{Values accepted: 'PreservedSpecimen' , 'LivingSpecimen', 'FossilSpecimen', 'HumanObservation', 'MachineObservation' or 'MaterialSample'.}

\item{CollectionCode}{Collection acronym to search. Ex: "FIOCRUZ-CEIOC", "UEC", "HUEFS".}

\item{CatalogNumber}{Register number of specimen in the collection.}

\item{Collector}{Collector name. Ex:    "Kock", "Almeida F".}

\item{CollectorNumber}{Collector number.}

\item{YearCollected}{Year of specimen was collected with four-digits. Accepts a vector.}

\item{IdentifiedBy}{Who identified the species of specimen. Ex: "Abreu MC".}

\item{YearIdentified}{Year of specimen was identified with four-digits.}

\item{Kingdom}{Kingdom classification name.}

\item{Phylum}{Phylum classification name.}

\item{Class}{Class classification name.}

\item{Order}{Order classification name.}

\item{Family}{Family classification name.}

\item{Genus}{Genus classification name.}

\item{ScientificName}{A character vector with scientific name, without authors. Accepts a vector of names. If you want include synonym taxa use \code{Synonyms} argument. See details to use 'embranco' or "naobranco" options.}

\item{TypeStatus}{.}

\item{Country}{Country of specimen register.}

\item{StateProvince}{State of specimen register.}

\item{County}{City of specimen register.}

\item{Locality}{Locality of specimen register. This field is very dificult to return with precision because it is a free form field.}

\item{MaxRecords}{numeric. The maximum number of general records to be returned. Default is to return all records. See details for more information.}

\item{all.data}{Logical. Output all data or coordinates only. FALSE is default.}

\item{Phonetic}{Logical. If TRUE, allows the system to ignore some spelling differences in scientific names like 'Y' or 'I', double letters, etc. Affects only taxonomic fields \code{Phylum}, \code{Class}, \code{Order}, \code{Family} e \code{ScientificName}. FALSE is default.}

\item{RedList}{Searches for records with scientific names that appear in red list of MMA if 'yes' or it that not appear in red list of MMA if 'no'. To return any records the values is NULL (default).}

\item{Scope}{Groups of organisms to search. The accepted values are "plants", "animals", "microrganisms" and "fossils". NULL to search in all groups (default).}

\item{Coordinates}{Choose if will download the registers with coordinates. Default is all registers (with coordinates or not). See details for available options.}

\item{Images}{If NULL (default) will search all record. If "yes" search the records that contain images of the specimen and 'no' will return only records without images. Images will not be downloaded. For more options see Details.}

\item{Synonyms}{Search also for synonyms defined in the following dictionaries: "sp2000", "flora2020", "MycoBank", "AlgaeBase", "DSMZ", "Moure" and/or NULL (default) if you don't want to search for synonyms.}

\item{Typus}{If 'yes' searches for type material only. If 'no', searches non type material only. Default is NULL and does no check type material.}
}
\value{
A list with parameters to use in \code{get_data()} function to download records.
}
\description{
This function prepare a list to request download of species register with the \code{get_data} function.
}
\details{
For more information about fields see \url{https://api.splink.org.br/} , \url{http://www.splink.org.br/} , \url{https://www.youtube.com/watch?v=I-4ftIWrKUA&feature=youtu.be} or \url{http://www.splink.org.br/showTips}.

The search system is case insensitive, meaning that searching for Tabeuia aurea is the same as searching for tabebuia aurea or TABEBUIA AUREA.

Accents are also not considered, meaning searching for São Paulo is the same as searching for Sao Paulo.

Avoid using words such as de, do, em, the, etc. For example when searching for species that occur in Serra da Canastra, search for Serra Canastra or just Canastra within the locality field.

In general, all words included in the form are considered when searching. Searching for Souza Silva in the collector’s field will retrieve all records that have the words Souza AND Silva simultaneously in that field.

\strong{Besides the connector |, users may search for a list of species – one in each line – with a maximum limit of 50 names.}

Users can also search for blank fields, by typing the word null in the field. For example, to search for unidentified material of the family Asteraceae, users must type Asteraceae in the family field and null in the scientific name field. Also for Asteracea, if you want to see what genera of the Asteraceae family have unidentified material, you may type notnull null in the scientific name field. The words null and notnull can be used in other fields.

The words 'null' and 'notnull' can also be used in the scientific name field. It is important to remember that the scientific name can be a genus (a single word), a binomial (genus + species) or a trinomial (genus + species + subspecies). The words 'null' and 'notnull' can replace any of these elements. For example: to search for records that have identification only up to the gender level, type 'notnull null' in the scientific name field.

Nos campos \strong{buscadores} são aceitos valores 'embranco' ou 'naobranco'. Por exemplo: é possível buscar apenas os registros de um gênero que não tenha sido identificado ainda, para isso deve-se informar no campo \code{ScientificName} "Manilkara embranco" (para o gênero Manilkara). Em caso de querer apenas os registros de um gênero que tenha identificação deve-se informar "Manilkara naobranco". No caso de querer todos os registros de um gênero, basta colocar "Manilkara".

Coordinates:  Yes | No | Original | Automatic | Blocked	  no check.

The complete list of searchable and modifiers fields are available with \code{fields_data()} (not implemented yet) function.

Attention, because the \code{MaxRecords} refers to the total number of records, in case of searching for more than one species it may be that records of only one species are returned. This depends on the number of records per species and the maximum number of records requested.

To the Images argument, the options are "yes" for records with images, "no" for records without images, "live" for records with images of living material, "pollen" for records with images of pollen, "wood" for images of wood (Xylotheque) and NULL to search all records.
}
\examples{
setup_list(ScientificName = "Manilkara maxima")

\dontrun{
setup_list(ScientificName = c("Rauvolfia sellowii", "Cantinoa althaeifolia"),
           StateProvince = c("São Paulo","Rio de Janeiro","Pernambuco"),
           MaxRecords = 5,
           Model = "Coords")
           }

}
